add readme / lib.rs
Move around some files to make cargo-rdme pick them up better
